Rename utils.rs to location.rs now that it only contains LineIndex
