+++ src/main.rs
@@ -1,11 +1,11 @@
 mod eval;
+mod location;
 mod nix_file;
 mod problem;
 mod ratchet;
 mod references;
 mod status;
 mod structure;
-mod utils;
 mod validation;
 
 use anyhow::Context as _;
